JitCore_x86: UDIV & UMOD return value and arguments sizes have to be the same

+++ miasm2/jitter/arch/JitCore_x86.h
@@ -275,18 +275,18 @@ void dump_gpregs(vm_cpu_t* vmcpu);
 uint64_t segm2addr(vm_cpu_t* vmcpu, uint64_t segm, uint64_t addr);
 
 
-uint32_t udiv64(vm_cpu_t* vmcpu, uint64_t a, uint64_t b);
-uint32_t umod64(vm_cpu_t* vmcpu, uint64_t a, uint64_t b);
+uint64_t udiv64(vm_cpu_t* vmcpu, uint64_t a, uint64_t b);
+uint64_t umod64(vm_cpu_t* vmcpu, uint64_t a, uint64_t b);
 int64_t idiv64(vm_cpu_t* vmcpu, int64_t a, int64_t b);
 int64_t imod64(vm_cpu_t* vmcpu, int64_t a, int64_t b);
 
-uint16_t udiv32(vm_cpu_t* vmcpu, uint32_t a, uint32_t b);
-uint16_t umod32(vm_cpu_t* vmcpu, uint32_t a, uint32_t b);
+uint32_t udiv32(vm_cpu_t* vmcpu, uint32_t a, uint32_t b);
+uint32_t umod32(vm_cpu_t* vmcpu, uint32_t a, uint32_t b);
 int32_t idiv32(vm_cpu_t* vmcpu, int32_t a, int32_t b);
 int32_t imod32(vm_cpu_t* vmcpu, int32_t a, int32_t b);
 
-uint8_t udiv16(vm_cpu_t* vmcpu, uint16_t a, uint16_t b);
-uint8_t umod16(vm_cpu_t* vmcpu, uint16_t a, uint16_t b);
+uint16_t udiv16(vm_cpu_t* vmcpu, uint16_t a, uint16_t b);
+uint16_t umod16(vm_cpu_t* vmcpu, uint16_t a, uint16_t b);
 int16_t idiv16(vm_cpu_t* vmcpu, int16_t a, int16_t b);
 int16_t imod16(vm_cpu_t* vmcpu, int16_t a, int16_t b);
